+++ Program.cs
@@ -1,6 +1,7 @@
 ﻿using LabManager.Database;
 using LabManager.Repositories;
 using Microsoft.Data.Sqlite;
+using LabManager.Models;
 
 var databaseConfig = new DatabaseConfig();
 var databaseSetup = new DatabaseSetup(databaseConfig);
@@ -29,17 +30,10 @@ if (modelName == "Computer")
         string ram = args[3];
         string processor = args[4];
 
-        var connection = new SqliteConnection("Data Source=database.db");
-        connection.Open();
+        var computer = new Computer(id, ram, processor);
+        computerRepository.Save(computer);
 
-        var command = connection.CreateCommand();
-        command.CommandText = "INSERT INTO Computers VALUES($id, $ram, $processor)";
-        command.Parameters.AddWithValue("$id", id);
-        command.Parameters.AddWithValue("$ram", ram);
-        command.Parameters.AddWithValue("$processor", processor);
-        command.ExecuteNonQuery();
 
-        connection.Close();
     }
 
 }
